Added replication tests
