+++ test/explain_test.rb
@@ -1,0 +1,41 @@
+require_relative "test_helper"
+
+class ExplainTest < Minitest::Test
+  def setup
+    City.delete_all
+  end
+
+  def test_explain
+    assert_match "Result", PgHero.explain("SELECT 1")
+  end
+
+  def test_explain_analyze
+    City.create!
+    assert_equal 1, City.count
+    PgHero.explain("ANALYZE DELETE FROM cities")
+    assert_equal 1, City.count
+  end
+
+  def test_explain_statement_timeout
+    with_explain_timeout(0.1) do
+      assert_raises(ActiveRecord::QueryCanceled) do
+        PgHero.explain("ANALYZE SELECT pg_sleep(1)")
+      end
+    end
+  end
+
+  def test_explain_multiple_statements
+    City.create!
+    assert_raises(ActiveRecord::StatementInvalid) { PgHero.explain("ANALYZE DELETE FROM cities; DELETE FROM cities; COMMIT") }
+  end
+
+  def with_explain_timeout(value)
+    previous_value = PgHero.explain_timeout_sec
+    begin
+      PgHero.explain_timeout_sec = value
+      yield
+    ensure
+      PgHero.explain_timeout_sec = previous_value
+    end
+  end
+end
